fix review's issues
