+++ test/data/vue-loader/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="utf-8">
-        <title>temp</title>
+        <title>vue-loader</title>
     </head>
     <body>
         <div id="app">
